:truck: Move Context to context.go
