+++ interfaces/handler/validator.go
@@ -3,26 +3,7 @@ package handler
 import (
 	"github.com/go-playground/validator/v10"
 	"github.com/gofrs/uuid"
-	"github.com/labstack/echo/v4"
-	"github.com/traPtitech/traPortfolio/usecases/repository"
 )
-
-type Context struct {
-	echo.Context
-}
-
-func (c *Context) BindAndValidate(i interface{}) error {
-	if err := c.Bind(i); err != nil {
-		c.Logger().Error(err)
-		return repository.ErrBind
-	}
-	if err := c.Validate(i); err != nil {
-		c.Logger().Error(err)
-		return repository.ErrValidate
-	}
-
-	return nil
-}
 
 type Validator struct {
 	validator *validator.Validate
